requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-vision
opencv-python-headless
ultralytics